+++ club/urls.py
@@ -6,5 +6,5 @@ from club.views import ClubIntroView
 app_name = 'club'
 
 urlpatterns = [
-    path('web/club-intro-web/', ClubIntroView.as_view(), name='list'),
+    path('web/club-intro-web/', ClubIntroView.as_view(), name='club-intro-web'),
 ]
