+++ search/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 from member.models import Member
-from teenplay_server.models import Period
+from teenplay_server.period import Period
 
 
 # class Search(Period):
@@ -16,10 +16,4 @@ class RecentSearch(Period):
         db_table = 'tbl_recent_search'
 
 
-class Like(Period):
-    member = models.ForeignKey(Member, null=False, blank=False, on_delete=models.PROTECT)
-    status = models.BooleanField(default=1, null=False, blank=False)
 
-    class Meta:
-        # 추상 모델을 설정(migrate 시 해당 모델의 테이블 생성X)
-        abstract = True
