+++ templates/club/web/base-web.html
@@ -7,7 +7,7 @@
           content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link href="https://cdn.jsdelivr.net/gh/sunn-us/SUIT/fonts/static/woff2/SUIT.css" rel="stylesheet" />
-    <link rel="shortcut icon" href="{% static 'staticfiles/images/teenplay_logo/logo1.png' %}" type="image/x-icon" />
+    <link rel="shortcut icon" href="{% static 'public/web/images/logo/logo1.png' %}" type="image/x-icon" />
     <link rel="stylesheet" href="{% static 'public/web/css/header-web.css' %}">
     <link rel="stylesheet" href="{% static 'public/web/css/footer-web.css' %}">
     {% block static %}{% endblock %}
